Add a module to get the latest Git tag
I added the module CMakeToolsVersionFromGit.cmake. The module
searches Git for the latest tag reachable from the current commit.
Then, it parses the tag information to create a version variable. The
project can use this variable for the project version. I also wrote the
user documentation for the new module.