Add support for MSI package
I fixed the installation destination to be correct on Windows. The Linux
package uses an absolute path, but WiX does not allow that. I set a
relative path on Windows.

I added support for the WiX package generator. I set some variables
specific to the WiX generator, and configured the license file so WiX
can use it in the package.